requirements.txt
adversarial_robustness_toolbox==1.13.1
easydict==1.10
numpy==1.23.4
matplotlib==3.7.1
seaborn==0.12.2
opencv_contrib_python==4.7.0.72
opencv_python==4.6.0.66
Pillow==9.5.0
python-dotenv==1.0.0
pytorch_lightning==1.9.4
scikit_learn==1.1.2
torch==2.0.0
torchmetrics==0.11.3
torchvision==0.15.1
tqdm==4.64.1
ipykernel==6.23.1
matplotlib_venn==0.11.9
ipywidgets==8.0.6